\name{spls.wrap-package}
\alias{spls.wrap-package}
\alias{spls.wrap}
\docType{package}
\title{
Wraps the spls package for variable selection and model-building.
}
\description{
Wraps the spls package for variable selection and model-building.
}
\details{
\tabular{ll}{
Package: \tab spls.wrap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-09-05\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Wesley Brooks

Maintainer: Wesley Brooks <wrbrooks@usgs.gov>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
